// js/stat-evaluation-player/src/generated/DoubleTapEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type DoubleTapEvent = { time: number, frame: number, player: RemoteIdTs, is_team_0: boolean, backboard_time: number, backboard_frame: number, };