// js/stat-evaluation-player/src/generated/DoubleTapPlayerStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Per-player accumulated double-tap stats.
 */
export type DoubleTapPlayerStats = { count: number, is_last_double_tap: boolean, last_double_tap_time: number | null, last_double_tap_frame: number | null, time_since_last_double_tap: number | null, frames_since_last_double_tap: number | null, };